Added rendering order. and I fixed a warning about loss of data in GameObject
git-svn-id: https://noobtoob-kickle.googlecode.com/svn/trunk@193 5e24083e-1465-11df-9d2a-2365f6df44e6

+++ Kickle/Kickle/Source Files/GameObject.cpp
@@ -421,7 +421,7 @@ bool GameObject::LoadObjectData( const std::string &filepath ) {
 
   // Load in speed of object
   std::string speed( root->FirstChildElement( "speed" )->GetText() );
-  m_speed = atof( speed.c_str() );
+  m_speed = (float)atof( speed.c_str() );
 
   // Load in whether to use rectangular collision for this object
   std::string collisionType = root->FirstChildElement( "collision_type" )->GetText();
